requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.8.1
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
multidict==5.2.0
notify==0.1.10
python-dateutil==2.8.2
python-twitch-client==0.7.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
twitchAPI==2.5.0
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validate-email==1.3
websockets==10.1
Werkzeug==2.0.2
yarl==1.7.2
gunicorn==19.7.1